Add history command

+++ src/commands/login.ts
@@ -2,7 +2,7 @@ import { Argument, Command } from "commander";
 import { setClientId, setClientSecret } from "../utils/auth";
 import { refreshToken } from "../utils/refreshToken";
 import { killProcess } from "../utils/childProcess";
-import { read, readFileSync } from "fs";
+import { exists, existsSync, readFileSync } from "fs";
 import path from "path";
 
 const clientId_arg = new Argument("[string]", "Client ID");
@@ -20,8 +20,10 @@ const login = new Command("login")
             await setClientId(clientId);
             await setClientSecret(clientSecret);
             const filePath = path.join(__dirname, "../utils/cur-pid.txt");
-            const pid = Number(readFileSync(filePath));
-            killProcess(pid);
+            if (existsSync(filePath)) {
+                const pid = Number(readFileSync(filePath));
+                killProcess(pid);
+            }
         }
 
         const opt = refreshToken();
